feat(BREAKING): add `absolute` option to config.getBuildPath
The second argument is now an options object, where before it was an optional string or buffer. To generate a content hash, define the `content` option.

The new `absolute` option can be set to false to ensure the returned file path is relative to the project. It can also be set to true for the opposite effect.